Add left-parenthesis to "URI special chars" class
This allows inputs like "email me (user@example.com)" and "the site
(http://example.com/)" to be correctly autolinked. Prior to this
change the left parenthesis was greedily matched as part of a
PRINTABLE token which then went on to consume the "http" or "user"
characters as well, preventing the recognition of a URI token.

Signed-off-by: Wincent Colaiuta <win@wincent.com>