Fix BLOCKQUOTE_START after lists
BLOCKQUOTE_START tokens were getting emitted as literal escapes if they
appeared after a list; this was because we explicitly prohibited them
from appearing inside scopes like LI. The problem then arises on leaving
such a list, because on the next line the scope will be "UL, LI" and so
the parser incorrectly forbids the BLOCKQUOTE_START in that context.

The solution is to refactor the BLOCKQUOTE_START case along similar
lines to PRE_START. We now explicitly disallow it inside BLOCKQUOTE
scopes, but everywhere else we employ the normal rollback-then-emit
behaviour. This produces correct output for well-formatted input, and
easily diagnosed bad output for malformed input.

Signed-off-by: Wincent Colaiuta <win@wincent.com>