Fix breakage caused by recent Rails API changes
The API has now changed so many times in unfathomable (uncommented) ways
that I'm no longer going to support older versions of Rails. The runtime
checks for all the myriad variations are just getting too complicated
and I can't justify maintaining them, especially seeing as I have no
intention of keeping old Rails installations on hand for the purposes of
testing.

The present version is tested and works with Rails 2.2.2. I actually had
a couple of intermediate, but uncommitted, unreleased changes, that I
needed to keep other versions of Rails working (2.1.0RC1, for example).
In the end those changes were short-lived because the API was changed
repeatedly afterwards.

In the latest rounds of changes, the "compilable?" method is no longer
checked, "call" is now invoked anyway, and Rails then proceeds to
evaluate the returned string as Ruby code in some unknown context where
"template" is a local variable! The "render" method is no longer called.

Signed-off-by: Wincent Colaiuta <win@wincent.com>