Initial set-up for tests against Rails
This isn't actually the final form in which I am likely to run these
tests but it lays down the basic prerequisites for actually being
able to test the extension against different versions of Rails.

Here we actually fire up a server to test the full stack; I suspect
that I nicer way of doing this will be to use Rails' built-in
integration tests instead.

Signed-off-by: Wincent Colaiuta <win@wincent.com>