Create "special links" from external rather than internal links
It was a clumsy design decision to offer detection of "special links"
inside internal links:

  [[/issues/20 | ticket #20]]

A much nicer syntax is to use external links instead:

  [/issues/20 ticket #20]

In this commit I rip out the old implementation as well as a
supporting instance variable (treat_slash_as_special), and a
couple of related low-level struct members used at the C
level (treat_slash_as_special and special_link) which were
themselves all probably evidence of "design smell".

In place of the old feature we move handling of "path"-style
links to inside of external links. Specs and docs are updated
accordingly.

See also:

  https://rails.wincent.com/issues/1208

Signed-off-by: Wincent Colaiuta <win@wincent.com>