Specify ":indent => false" default in wikitext/string extension
Seeing as the String extension is primarily for use in Rails
applications, where setting up Haml to run with "ugly" mode turned
on is a good idea, it makes sense to make the "w" and "to_wikitext"
methods on the String class pass in ":indent => false" by default.

This can be overridden if desired by passing in an explicit indent
such as ":indent => 0".

See:

  https://wincent.com/issues/817

Signed-off-by: Wincent Colaiuta <win@wincent.com>