Preliminary fix for Ruby 1.9 compatibility
The migration to the str_t type for the output buffer broke
compatibility with Ruby 1.9.

This commit shows what things would look like if we could
do a compile-time check for 1.9 and could thus modify the code
to make it compatible.

Unfortunately this doesn't actually work because although Ruby
comes with a "version.h" file with all the necessary macros,
for some reason it isn't installed when doing "make install".

In my local testing, that means that I can build under 1.9,
but it will fall back to the system-installed (1.8) headers
and the values for RUBY_VERSION_* are not correct.

Signed-off-by: Wincent Colaiuta <win@wincent.com>