Change internal function prefix from "_Wikitext_" to "wiki_"
Follow the pattern already established in str.c/str.h and
ary.c/ary.h, and use a short lowercase prefix with no
leading underscore for "internal" functions (that is,
functions that are for use within the C part of the extension
and are not exposed as Ruby methods).

This buys us some horizontal space and makes keeping things
in a reasonable number of columns a bit easier.

Externally exported methods still follow the old pattern of
Module_class_method (eg. Wikitext_parser_parse).

Signed-off-by: Wincent Colaiuta <win@wincent.com>