Fix builds on 64-bit Darwin platforms
Way back in ancient history when wikitext still used an ANTLR-generated
lexer, the evil commit 7a510102 added a horrible specific ARCHFLAGS
override to overcome the fact that ANTLR didn't produce Universal
binaries.

This worked fine on 32-bit machines, and the evil commit was allowed to
persist through the codebase for over two years, long after ANTLR was
ripped out and replaced with the vastly superior Ragel-generated lexer.

Today on testing on a 64-bit Darwin machine for the first time I
discovered the evil, and this commit expurgates it.

Signed-off-by: Wincent Colaiuta <win@wincent.com>