Fix CSS class of explicit mailto: URIs
Previously we only correctly handled "mailto" links when the link target
itself included the "mailto:" prefix.

In other cases, where the prefix was passed in separately, for example,
we did not apply the special CSS class.

Now we always check both the link target and the prefix to see if either
of them begins with "mailto:".

Fixes the spec failures reported in:

  https://wincent.com/issues/1389

Signed-off-by: Wincent Colaiuta <win@wincent.com>