Extract gemspec out into separate file
As recommended here for compatibility with Bundler's "install from Git
repo" feature:

  http://yehudakatz.com/2010/04/02/using-gemspecs-as-intended/

Note that while there is some disagreement in the comments on that post
about whether this is a clever or correct thing to do, in the present
case we are merely extracting the hand-edited gemspec out into a
separate file, not "checking in a generated file into the repo".

The change seems justified because it buys us compatibility with a
Bundler feature at no real cost.

Likewise, the other main complaint raised on a different post on the
same blog (ie. that building from Git repos can run into complications
where you end up with two different gems with different code but the
same version number) doesn't apply here because I have always used a
sane version numbering scheme, bumping the version number after each
release, so a released gem will never have the same version as an
intermediate version built directly from the repo:

  http://yehudakatz.com/2010/04/12/some-of-the-problems-bundler-solves/

Signed-off-by: Wincent Colaiuta <win@wincent.com>