Rakefile: run "make" target before "gem build"
This is the reincarnation of commit 829011d, which fixed a subtle
problem with a missing file in the built gem.

We basically have to make sure that we've run the "make" target so that
all of the C files, including the Ragel-generated C file, is present in
the directory at the time "gem build" is called.

Signed-off-by: Wincent Colaiuta <win@wincent.com>