Split String#wikitext_preprocess into a separate file
This has always been a bit of a wart, the fact that what is essentially
an application-specific preprocessing pass was included in our
"wikitext/strings" file.

Now, if you want the behavior you have to explicitly require
'wikitext/preprocess'.

As a side effect of this change, the file now serves as a nice example
of how to implement a custom preprocessing phase if you need it.

Signed-off-by: Wincent Colaiuta <win@wincent.com>