Run specs under Rails 3.0.0.beta4
Note how that when targetting Rails 3 release versions, we can't use the
version number as the app name as we did with Rails 2.

This is because Rails 3 tries to turn the name into a constant, and
constants can't start with a number. To work around this we take such
version numbers and prepend them with a "v" to create a test app name;
this in turn can be "constantized" into "V300Beta4" or similar.

Signed-off-by: Wincent Colaiuta <win@wincent.com>