Backend support for HTML5/XML output styles
HTML5 has two output syntaxes, HTML and XML. Under HTML5 the HTML output
syntax is likely to be predominant, given that the XML syntax can only
be used if it is served with an appropriate MIME type (ie.
application/xml), and the vast majority of deployed web servers are
configured to use a text/html MIME type.

So, this dictates that for our 2.0 release we are going to default to
emitting HTML syntax, but allow the user to override the default and
emit XML syntax if required.

In practice, the only place in the code base where this necessitates a
change is the part where we emit "img" tags. In the XML syntax, these
tags should be self-closing; ie:

  <img src="foo.png" alt="Foo!" />

Whereas in the HTML syntax they should not (even if all user agents
accept them anyway); ie:

  <img src="foo.png" alt="Foo!">

Signed-off-by: Wincent Colaiuta <win@wincent.com>