Format all "pre" blocks for YARD
This means marking all non-Ruby "pre" blocks using a "!!!wikitext" or
"!!!html" marker.

At the moment there is no syntax highlighting for wikitext or HTML, so
those tags are effectively equivalent to "!!!plain", but the important
thing is that we stop YARD from inappropriately applying Ruby syntax
highlighting rules to non-ruby code, and if syntax highlighting for the
other languages is ever added, we will be ready for it.

Signed-off-by: Wincent Colaiuta <win@wincent.com>