Use Bundler to manage development environment
Note that this actually breaks many (all?) of the specs which run
in the scratch Rails environments set up to test the extension against
different versions of Rails.

This is because Bundler tries to apply the same restrictions to those
sub-environments as apply to the parent environment, so more changes
will be needed to adapt those Rails tests to work in a Bundler-dominated
world.

Signed-off-by: Wincent Colaiuta <win@wincent.com>