Teach wiki_append_sanitized_link_target to take a str_t
Rather than passing in a parser_t and assuming the text we want to
operate on is the one in the struct, pass in the link target str_t
explicitly. This added flexibility will make the
wiki_append_sanitized_link_target function suitable for wider use, which
I am presently going to do in order to fix a bug.

Signed-off-by: Wincent Colaiuta <win@wincent.com>