chore: mark 3.1.0 as supported
At this point, I am sick of having to update this on a yearly basis. For
examples of previous updates, see:

- 2.3.0: https://github.com/wincent/wikitext/issues/3
- 2.4.0: https://github.com/wincent/wikitext/issues/4

As noted on those threads, given that wikitext does some low-level
struct stuff in C-land in the name of speed, I used to want to manually
verify that it still worked with each version. But it is over 14 years
now since I first imported this project into Git, and it has existed for
even longer than that, and I don't recall the struct format every
breaking during that time. If it does break, I expect it will break in
an obvious and appalling way that will be no more obvious and appalling
than failing to build.

So, bump to 3.1.0 in development, and use Frum to manage my local ruby
version:

    https://github.com/tako8ki/frum

Then:

    eval "$(frum init)"
    frum install 3.1.0
    frum local 3.1.0

Adjust the tests to make the one failing test work. And let's call this
good.